Move rating modal to pre logout
