+++ Apps/WebClient/src/ClientApp/src/components/navmenu/navHeader.vue
@@ -54,33 +54,36 @@ nav {
 
         <!-- Navbar links -->
         <b-navbar-nav class="ml-auto">
-            <router-link
+            <b-btn
                 v-if="oidcIsAuthenticated"
-                id="menuBtnLogin"
+                id="menuBtnLogout"
+                variant="link"
                 class="nav-link"
-                to="/logout"
+                @click="showRating()"
             >
                 <font-awesome-icon icon="sign-out-alt"></font-awesome-icon>
                 Logout
-            </router-link>
+            </b-btn>
             <router-link v-else id="menuBtnLogin" class="nav-link" to="/login">
                 <font-awesome-icon icon="sign-in-alt"></font-awesome-icon> Login
             </router-link>
         </b-navbar-nav>
+        <RatingComponent ref="ratingComponent" :on-close="modalClosed()" />
     </b-navbar>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
-import { Component, Prop, Watch } from "vue-property-decorator";
+import { Component, Prop, Watch, Ref } from "vue-property-decorator";
 import { Action, Getter } from "vuex-class";
 import { User as OidcUser } from "oidc-client";
-import { IAuthenticationService } from "@/services/interfaces";
+import { ILogger } from "@/services/interfaces";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
 import container from "@/plugins/inversify.config";
 import User from "@/models/user";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { faSignInAlt, faSignOutAlt } from "@fortawesome/free-solid-svg-icons";
+import RatingComponent from "@/components/modal/rating.vue";
 library.add(faSignInAlt);
 library.add(faSignOutAlt);
 
@@ -93,8 +96,13 @@ const auth: string = "auth";
 const user: string = "user";
 const sidebar: string = "sidebar";
 
-@Component
+@Component({
+    components: {
+        RatingComponent: RatingComponent,
+    },
+})
 export default class HeaderComponent extends Vue {
+    private logger: ILogger = container.get(SERVICE_IDENTIFIER.Logger);
     @Action("toggleSidebar", { namespace: sidebar }) toggleSidebar!: () => void;
     @Getter("isOpen", { namespace: sidebar }) isOpen!: boolean;
 
@@ -111,13 +119,8 @@ export default class HeaderComponent extends Vue {
     @Getter("userIsActive", { namespace: user })
     userIsActive!: boolean;
 
-    private authenticationService!: IAuthenticationService;
-
-    private mounted() {
-        this.authenticationService = container.get(
-            SERVICE_IDENTIFIER.AuthenticationService
-        );
-    }
+    @Ref("ratingComponent")
+    readonly ratingComponent!: RatingComponent;
 
     private get displayMenu(): boolean {
         return (
@@ -129,6 +132,15 @@ export default class HeaderComponent extends Vue {
 
     private toggleMenu() {
         this.toggleSidebar();
+    }
+
+    private showRating() {
+        this.ratingComponent.showModal();
+    }
+
+    private modalClosed() {
+        this.logger.debug(`redirecting to logout view ...`);
+        this.$router.push({ path: "/logout" });
     }
 }
 </script>
